+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Not released
 
+## 1.3
+
+### 1.3.0-alpha.12 (2022-06-01)
+
 - Tiles features calculation for spatial indexes [#417](https://github.com/CartoDB/carto-react/pull/417)
 - Add layer options extension mechanism in LegendWidget [#416](https://github.com/CartoDB/carto-react/pull/416) 
 - Use new Query Models in SQL API for global widgets [#415](https://github.com/CartoDB/carto-react/pull/415)
-
-## 1.3
 
 ### 1.3.0-alpha.11 (2022-05-20)